Squashed commit of the following:
commit aadf576c236b137169aa5f4200f74028082021cb
Author: Abdelhamid Bakhta <abdelhamid.bakhta@gmail.com>
Date:   Mon Nov 21 17:58:23 2022 +0100

    🚚 rename package name

commit 2731233a268d293fdf956e77b1bb424b3bc02bed
Author: Abdelhamid Bakhta <abdelhamid.bakhta@gmail.com>
Date:   Mon Nov 21 17:56:41 2022 +0100

    🔥 remove pr labels workflow

commit 84d6cdd8c1eb5921e2345be6df166cfe890fc3ec
Author: Abdelhamid Bakhta <abdelhamid.bakhta@gmail.com>
Date:   Mon Nov 21 17:54:48 2022 +0100

    ✨ switch to workspace

commit 9bc7e007e0e9f9d0ed536213ea113718390caa31
Author: Abdelhamid Bakhta <abdelhamid.bakhta@gmail.com>
Date:   Mon Nov 21 15:58:48 2022 +0100

    🚧 switch to using workspace
