+++ cli/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "cli"
+version = "0.1.0"
+edition.workspace = true
+description.workspace = true
+homepage.workspace = true
+
+[dependencies]
+core = { path = "../core" }
+helios = { git = "https://github.com/a16z/helios", rev = "2dbe057e"}
+ethers ={ git = "https://github.com/gakonst/ethers-rs", rev = "5c9a048f"}
+ethers-etherscan = "1.0.0"
+primitive-types = "0.11.1"
+revm = "2.1.0"
+eyre = "0.6.8"
+clap = { version = "4.0.22", features = ["derive"] }
+serde = { version = "1.0.147", features = ["derive"] }
+serde_json = "1.0"
+tokio = { version = "1.21.2", features = ["full"] }
