--- src/cli/mod.rs
@@ -1,3 +1,0 @@
-pub mod ethereum;
-pub mod model;
-pub mod runner;
